requirements.txt
asgiref==3.2.7
bson==0.5.8
dataclasses==0.6
Django==2.2.12
djangorestframework==3.11.0
djongo==1.3.2
elasticsearch==7.6.0
elasticsearch-dsl==7.1.0
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.2.4
urllib3==1.25.9
